+++ flutter/lib/src/navigation/sentry_navigator_observer.dart
@@ -6,7 +6,7 @@ import 'package:flutter/widgets.dart';
 
 import '../../sentry_flutter.dart';
 import '../sentry_native.dart';
-import '../sentry_native_wrapper.dart';
+import '../sentry_native_channel.dart';
 
 /// This key must be used so that the web interface displays the events nicely
 /// See https://develop.sentry.dev/sdk/event-payloads/breadcrumbs/
