+++ flutter/lib/src/sentry_flutter.dart
@@ -6,7 +6,7 @@ import 'package:meta/meta.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:sentry/sentry.dart';
 import 'sentry_native.dart';
-import 'sentry_native_wrapper.dart';
+import 'sentry_native_channel.dart';
 
 import 'flutter_enricher_event_processor.dart';
 import 'integrations/debug_print_integration.dart';
