+++ flutter/lib/src/sentry_native_channel.dart
