+++ flutter/test/integrations/native_app_start_integration_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/integrations/native_app_start_integration.dart';
 import 'package:sentry_flutter/src/sentry_native.dart';
-import 'package:sentry_flutter/src/sentry_native_wrapper.dart';
+import 'package:sentry_flutter/src/sentry_native_channel.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 import 'package:flutter/scheduler.dart';
